+++ src/application/application.js
@@ -7,7 +7,7 @@ import timer from "./../system/timer.js";
 import state from "./../state/state.js";
 import World from "./../physics/world.js";
 import { onresize } from "./resize.js";
-import { defaultSettings } from "./settings.js";
+import { ApplicationSettings } from "./settings.js";
 import { consoleHeader } from "./header.js";
 import { CANVAS, WEBGL, AUTO } from "../const.js";
 
@@ -20,7 +20,7 @@ export default class Application {
     /**
      * @param {number} width - The width of the canvas viewport
      * @param {number} height - The height of the canvas viewport
-     * @param {Application.Settings} [options] - The optional parameters for the application and default renderer
+     * @param {ApplicationSettings} [options] - The optional parameters for the application and default renderer
      * @throws Will throw an exception if it fails to instantiate a renderer
      */
     constructor(width, height, options) {
@@ -76,7 +76,7 @@ export default class Application {
 
         /**
          * the given settings used when creating this application
-         * @type {Application.settings}
+         * @type {ApplicationSettings}
          */
         this.settings = undefined;
 
@@ -137,11 +137,11 @@ export default class Application {
      * init the game instance (create a physic world, update starting time, etc..)
      * @param {number} width - The width of the canvas viewport
      * @param {number} height - The height of the canvas viewport
-     * @param {Application.Settings} [options] - The optional parameters for the application and default renderer
+     * @param {ApplicationSettings} [options] - The optional parameters for the application and default renderer
      */
     init(width, height, options) {
 
-        this.settings = Object.assign(defaultSettings, options || {});
+        this.settings = Object.assign(ApplicationSettings, options || {});
 
         // sanitize potential given parameters
         this.settings.width = width;