+++ src/video/video.js
@@ -2,6 +2,7 @@ import * as event from "./../system/event.js";
 import { initialized, game } from "./../index.js";
 import * as device from "./../system/device.js";
 import { throttle } from "../utils/function.js";
+import { ApplicationSettings } from "../application/settings.js";
 
 /**
  * @namespace video
@@ -43,7 +44,7 @@ export let renderer = null;
  * @memberof video
  * @param {number} width - The width of the canvas viewport
  * @param {number} height - The height of the canvas viewport
- * @param {Application.Settings} [options] - optional parameters for the renderer
+ * @param {ApplicationSettings} [options] - optional parameters for the renderer
  * @returns {boolean} false if initialization failed (canvas not supported)
  * @example
  * // init the video with a 640x480 canvas
@@ -62,7 +63,7 @@ export function init(width, height, options) {
 
     try {
         // initialize the default game Application with the given options
-        game.init(width, height, options);
+        game.init(width, height, Object.assign(ApplicationSettings, options || {}));
     } catch (e) {
         console.log(e.message);
         // me.video.init() historically returns false if failing at creating/using a HTML5 canvas